requirements.txt
psutil>=5.9.0
pynput>=1.7.0
pywin32>=305; sys_platform == 'win32'
pystray>=0.19.0
Pillow>=9.0.0
opencv-python>=4.8.0
pyautogui>=0.9.54
numpy>=1.24.0